feat ✨: Implemented CRUD functionality and refactored database management with custom instance
- Implemented CRUD functionality for employee management.
- Added CRUD operations for item categories.
- Implemented database initialization and connection setup with ORM registration.
- Implemented database connection initialization using Beego ORM.
- Refactored application to use a custom database instance and removed Beego's ORM for database management.
- Changed data type of IdCondition field from uint8 to uint in ConditionLookup struct.
- Refactored DocUploaded struct to simplify relationships and columns.
- Refactored Item model to correctly define Category field as a foreign key to ItemCategory.
- Refactored ItemUnit model to improve relationships and column specifications.
- Refactored Lending struct to use pointer references for foreign key relationships.
- Updated data type of IdType field in LendingTypeLookup struct to accommodate larger values.
- Changed data type of IdRepT field in RepTypeLookup struct from int8 to uint.
- "Updated RepairLog struct to use pointers for foreign key fields and added Unit field to enable one-to-one relationship with ItemUnit."
- Changed data type of IdStatus field in StatusLookup struct from uint8 to uint.
- Refactored User model to improve data relationships and consistency.
- Added new routes for item units and refactored the router initialization to include database setup.

+++ models/condition_lookup.go
@@ -5,7 +5,7 @@ import (
 )
 
 type ConditionLookup struct {
-    IdCondition    uint8  `orm:"column(id_condition);auto;pk" description:"primary key"`
+    IdCondition    uint  `orm:"column(id_condition);auto;pk" description:"primary key"`
     ConditionName  string `orm:"column(condition_name);size(255);null(false)"`
 }
