+++ models/status_lookup.go
@@ -5,7 +5,7 @@ import (
 )
 
 type StatusLookup struct {
-    IdStatus    uint8  `orm:"column(id_status);auto;pk" description:"primary key"`
+    IdStatus    uint  `orm:"column(id_status);auto;pk" description:"primary key"`
     StatusName  string `orm:"column(status_name);size(30);null(false)" description:"status of the unit"`
 }
